Update Android build to use multi-project setup
Use multi-project build setup instead of build flavours
for samples. The root project `build.gradle` specifies
how to build, and each subproject just creates a copyTask
to copy assets that are used by the project.

Updates/benefits:
- In Android Studio, easier way to find and switch samples
and no gradle sync when switching build flavours
- In CLI more straightforward way to discover projects and
no need to remember syntax on how to build.
- Specifies assets per project, resulting in smaller APKs.
Third party assets are only used in a few projects, and
those are very large (> 300 MB). No need to add them to
projects that don't use them.
- Instead of split APKS use abi filters to only build
for arm64-v8a
- Build all samples in the github CI

One drawback of the new setup is specifying assets
that are used inside the project, so if the project is
updated we have to remember to update the build.gradle file,
and there is also no automated way to test right now.